blogicum/requirements.txt
asgiref==3.8.1
attrs==24.3.0
Django==5.1.5
django-debug-toolbar==5.0.1
flake8==5.0.4
flake8-docstrings==1.7.0
iniconfig==2.0.0
mccabe==0.7.0
packaging==24.2
pep8-naming==0.13.3
pluggy==1.5.0
py==1.11.0
pycodestyle==2.9.1
pydocstyle==6.3.0
pyflakes==2.5.0
pytest==7.1.3
pytest-django==4.5.2
pytest-pythonpath==0.7.3
pytz==2024.2
snowballstemmer==2.2.0
sqlparse==0.5.3
tomli==2.2.1
